Force sequential steps for promise resolution when starting recording in the example Chat app.
Was getting the following error on occasion:

```
DOMException: Failed to execute 'createAnswer' on 'RTCPeerConnection': PeerConnection cannot create an answer in a state other than have-remote-offer or have-local-pranswer.
```

This seems to prevent it. The await on the promise list isn't required per-se but keeping it retains behavior similar to pre-patch in case there are other parts of the code that rely on this function waiting until the recording is ready before proceeding.